Fix crash on amd64 when enabling file logging
On amd64 architectures, slirp crashes immediately when file logging is
enabled with "log start".  It happens at the first call to lprint after the
log file is opened and ready for logging.  The reason is that the va_list
object "args" is used twice in that case, once at

    lprint_ptr += (*lprint_print)(*lprint_arg, format, args);

and once at

    vfprintf(lfd, bptr2, args);

After the first call, args gets invalid because all arguments have been
read from the va_args structure.  For some reason, this does not happen on
i386 versions. Maybe a real copy of args is created when the function is
called in i386.