Do not run MAKEPRO anymore
mkpro is a nice perl script that used to regenerate all slirp prototypes
correctly in the past. It stopped working, maybe because a change in perl
and/or because it has a bug. I'm not a perl guru, and after wasting hours
of my time trying to decypher mkrpo and fixing that silly thing, I quit.
Slirp hasn't changed for years, so I think mkpro isn't very usefull anyway.